Create foods and meals directories in test directory to keep test files organized
