+++ test/foods/get_show_request.spec.js
@@ -2,7 +2,7 @@ const pry = require('pryjs')
 const chai = require('chai');
 const should = chai.should();
 const chaiHttp = require('chai-http');
-const server = require('../server');
+const server = require('../../server');
 
 chai.use(chaiHttp);
