+++ test/foods/post_request.spec.js
@@ -1,7 +1,7 @@
 const chai = require('chai');
 const should = chai.should();
 const chaiHttp = require('chai-http');
-const server = require('../server');
+const server = require('../../server');
 
 chai.use(chaiHttp);
